ref(billing): Simplify localStorage tests and fix isolation bug
Consolidate 3 redundant unit tests into 1 focused test that covers
render, auto-open, and setItem gracefully. Add afterEach with
jest.restoreAllMocks() to prevent mockReturnValue leaking across
describe blocks (clearMocks only resets call history, not implementations).

Move createStorage utility tests to a dedicated createStorage.spec.tsx
file — they test shared utility behavior, not component behavior.
Remove implementation-detail test that only verified call arguments.
